+++ src/Concerto/TestBundle/Resources/R/service.R
@@ -4,7 +4,7 @@ ENV_CONCERTO_R_FORCED_GC_INTERVAL = as.numeric(Sys.getenv("CONCERTO_R_FORCED_GC_
 concerto.log("starting service listener")
 
 queue = c()
-unlink(paste0(ENV_CONCERTO_R_SERVICE_FIFO_PATH, "*"))
+#unlink(paste0(ENV_CONCERTO_R_SERVICE_FIFO_PATH, "*"))
 lastForcedGcTime = as.numeric(Sys.time())
 repeat {
   if(ENV_CONCERTO_R_FORCED_GC_INTERVAL >= 0) {
@@ -17,11 +17,18 @@ repeat {
 
   reqFifoPath = NULL
   if(length(queue) == 0) {
-    queue = list.files(ENV_CONCERTO_R_SERVICE_FIFO_PATH, pattern=".*\\.reqfifo", full.names=TRUE)
+    queue = list.files(ENV_CONCERTO_R_SERVICE_FIFO_PATH, pattern=".*\\.reqfifo$", full.names=TRUE)
   }
+
+  fl = NULL
+  lockPath = NULL
   if(length(queue) > 0) {
     reqFifoPath = queue[1]
     queue = queue[-1]
+
+    lockPath = paste0(reqFifoPath, ".lock")
+    fl = lock(lockPath, exclusive=T, timeout=0)
+    if(is.null(fl)) { next }
   } else {
     Sys.sleep(0.01)
     next
@@ -31,6 +38,8 @@ repeat {
   request = readLines(con)
   close(con)
   unlink(reqFifoPath)
+  unlock(fl)
+  unlink(lockPath)
 
   request = paste0(request, collapse="\n")
   request = charToRaw(request)