+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * fixed forker guard cron job check
 * fixed files permissions when importing content
+* client side error logging should now log all js exceptions
+* fixed 403 on client side error logging
 
 #### Improvements