+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fixed files permissions when importing content
 * client side error logging should now log all js exceptions
 * fixed 403 on client side error logging
+* fixed 'options' item type fixed index
 
 #### Improvements