+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * updated R to v4.0
 * exposed test_runner_settings.r_profile_session_path as env var
 * exposed test_runner_settings.r_environ_session_path as env var
-* added CONCERTO_R_SERVICE env var
+* added CONCERTO_R_SERVICES_NUM env var
 
 ## 5.0.22 (in progress)