+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * configurable R long running processes forced GC interval (r_forced_gc_interval)
 * replaced RSA key pair for JWT with HMAC key (CONCERTO_JWT_SECRET env)
+* exposed testRunner.getControlValues()
 
 #### Starter content